moved base_map_codec and fixed docs about codecs
