+++ pumpkin-datafixerupper/src/serialization/map_codecs/mod.rs
@@ -1,4 +1,3 @@
-pub mod base;
 pub mod field_coders;
 pub mod optional_field;
 pub mod simple;
